+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changes for Iris Admin UI
 Unreleased
 ==========
 
-- Petition Editor: generate appendix
+- Petition Editor: generate letter cover and appendix
 - Petition Editor: reset og
 - Petition Editor: fix og url
 - Petition Filter: wording of filter / sort